Fix an issue causing messages from the launcher to be lost.
Since the launcher sends the message data and the ending newline via
separate calls to send, it's possible for the game to receive a
message's data but not the following newline. This is considered a
timeout and the partial message data is returned in the partial
result--but this was being ignored, resulting in the message being
effectively lost.

This seems to be a much more prevalent issue when running the client
natively on Linux, and seems to be more noticeable in MPGameNetwork,
resulting in vehicle events, chat messages, server-triggered events, and
more being silently dropped.